Fix: 에러 로그를 500번대 예외와 exception에 대해서만 db에 남기도록 변경
