+++ src/main/java/smash/teams/be/core/advice/LogAdvice.java
@@ -8,12 +8,7 @@ import org.aspectj.lang.annotation.Aspect;
 import org.aspectj.lang.annotation.Before;
 import org.aspectj.lang.annotation.Pointcut;
 import org.aspectj.lang.reflect.MethodSignature;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
-import smash.teams.be.core.auth.session.MyUserDetails;
-import smash.teams.be.model.errorLog.ErrorLog;
-import smash.teams.be.model.errorLog.ErrorLogRepository;
 
 import java.lang.reflect.Method;
 
@@ -22,8 +17,6 @@ import java.lang.reflect.Method;
 @Aspect
 @Component
 public class LogAdvice {
-
-    private final ErrorLogRepository errorLogRepository;
 
     @Pointcut("@annotation(smash.teams.be.core.annotation.Log)")
     public void log() {
@@ -48,25 +41,7 @@ public class LogAdvice {
             if (arg instanceof Exception) {
                 Exception e = (Exception) arg;
                 log.error("에러 : " + e.getMessage());
-                errorLogRepository.save(ErrorLog.builder()
-                        .msg(e.getMessage())
-                        .userId(getUserId())
-                        .build());
             }
         }
-    }
-
-    private Long getUserId() {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        if (authentication == null || !authentication.isAuthenticated()) {
-            return null;
-        }
-
-        Object principal = authentication.getPrincipal();
-        if (principal instanceof MyUserDetails) {
-            return ((MyUserDetails) principal).getUser().getId();
-        }
-
-        return null;
     }
 }
